Feat: Completed the integration of AutoMapper and FluentValidation. Also improved code clarity and did some clean-up.
In this update, I finished adding AutoMapper and FluentValidation into the project. I created the necessary mapping profiles for the main models and tested them to make sure the mappings work correctly. I also added the validation rules by using FluentValidation, so now the incoming data is checked in a more structured and readable way.

During this work, I noticed that some parts of the code were a bit confusing or repeated, so I did some small refactoring to make it easier to understand. I removed unused code blocks, renamed a few variables to more meaningful names, and organized the files to follow a clearer structure. I think this will help the project be easier to maintain in the future.

Overall, this commit improves the code quality and makes the project more consistent by using both AutoMapper and FluentValidation together.